[COMMON] mm: gup: fix invalid comparison for CMA
CMA pages should not be pinned for a long time because they should be
always ready for reclamation by the user's request. If a user of gup
wants to pin a series of pages for a long time he/she should give
FOLL_CMA to let gup migrate the CMA pages that are in pinning to
non-CMA pages.

In order to determine if a pinning page is in a CMA page, gpu should
compare if the migrate type of the pageblock of the page is
MIGRATE_CMA. But the comparison is not correct and the migration
never happens.

SVACE #82317

Change-Id: Id4ef40957365bdbe586f29c4ba0e6318e232719a
Signed-off-by: Cho KyongHo <pullip.cho@samsung.com>