media: mfc: Fix the reference count pair.
When decoding and display status,
the release information of the referenced buffers without display should not be updated.
It will be reused.

When decoding only status,
the release information of the referenced buffers(except for reused buffer)
should be left to update when displayed.
It is an already displayed buffer.

Change-Id: I37df2bfd9d6cd3defde26edf1f2ac46ebb42ae8e
Signed-off-by: Jeonghee Kim <jhhhh.kim@samsung.com>